+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/app/ui.js
@@ -5,9 +5,40 @@ UiController = Marionette.Controller.extend({
     initialize: function() {
         log('UiController.initialize');
     },
+
     setup_text_tools: function() {
         // auto-shorten some texts
         $(".very-short").shorten({showChars: 5, moreText: 'more', lessText: 'less'});
+    },
+
+    notify: function(message, options) {
+
+        if (options.icon) {
+            message = '<span class="icon ' + options.icon + ' icon-large"></span>' + '<p>' + message + '</p>';
+        }
+        if (!options.type) {
+            options.type = 'notice';
+        }
+
+        //setTimeout( function() {
+
+            // create the notification
+            var notification = new NotificationFx({
+                message : message,
+                layout : 'attached',
+                effect : 'bouncyflip',
+                type :   options.type, // notice, warning, error, success
+                onClose : function() {
+                    //bttn.disabled = false;
+                },
+                //ttl: 100000,
+            });
+
+            // show the notification
+            notification.show();
+
+        //}, 1200 );
+
     },
 });